MOD rename project and packages
