+++ .idea/compiler.xml
@@ -6,11 +6,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="javafx-base" />
+        <module name="elevator-control-center" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
-      <module name="javafx-base" target="11" />
+      <module name="elevator-control-center" target="11" />
     </bytecodeTargetLevel>
   </component>
 </project>
