+++ pom.xml
@@ -1,9 +1,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>at.fhhagenberg.sqe</groupId>
-    <artifactId>javafx-base</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <groupId>at.fhhgb.team.a.elevators</groupId>
+    <artifactId>elevator-control-center</artifactId>
+    <version>0.0.1</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -64,7 +64,7 @@
                 <version>0.0.3</version>
                 <configuration>
                     <launcher>launcher</launcher>
-                    <mainClass>at.fhhagenberg.sqe.App</mainClass>
+                    <mainClass>at.fhhgb.team.a.elevators.App</mainClass>
                 </configuration>
             </plugin>
             <plugin>
